Update StorageManager to include static map blobs

+++ Actively/BlobStorage/StorageManager.cs
@@ -6,33 +6,69 @@ namespace Actively.BlobStorage
 	public class StorageManager : IStorageManager
 	{
 		private readonly string _connectionString;
-		private const string _geojsonRoutesContainerName = "geojson-routes";
+		private readonly Dictionary<BlobType, string> _fileExtensions = new()
+		{
+			{BlobType.Geojson, ".geojson" },
+			{BlobType.StaticMap, ".png" }
+		};
+		private readonly Dictionary<BlobType, string> _containerNames = new()
+		{
+			{BlobType.Geojson, "geojson-routes" },
+			{BlobType.StaticMap, "static-maps" }
+		};
 
 		public StorageManager(IConfiguration configuration)
 		{
 			_connectionString = configuration.GetSection("AzureBlob").Value!;
 		}
 
-		private BlobClient CreateBlob(Guid activityId)
+		private BlobClient CreateBlob(Guid activityId, BlobType type)
 		{
-			var blobName = activityId + ".geojson";
-			var containerClient = new BlobContainerClient(_connectionString, _geojsonRoutesContainerName);
+			if(!_fileExtensions.TryGetValue(type, out var fileExtension))
+			{
+				throw new ArgumentException("Provided blob type is invalid");
+			}
+
+			if(!_containerNames.TryGetValue(type, out var containerName))
+			{
+				throw new ArgumentException("Provided blob type is invalid");
+			}
+
+			var blobName = activityId + fileExtension;
+			var containerClient = new BlobContainerClient(_connectionString, containerName);
 			containerClient.CreateIfNotExistsAsync();
 			return containerClient.GetBlobClient(blobName);
 
 		}
 
-		public async Task Upload(Guid activityId, Stream content)
+		public async Task Upload(Guid activityId, BlobType type, Stream content)
 		{
-			var blob = CreateBlob(activityId);
+			var blob = CreateBlob(activityId, type);
 			content.Position = 0;
 			await blob.UploadAsync(content);
 		}
 
-		public async Task Delete(Guid activityId)
+		//deletes all blobs related to activity with given activityId
+		public async Task DeleteActivityBlobs(Guid activityId)
 		{
-			var blobName = activityId + ".geojson";
-			var containerClient = new BlobContainerClient(_connectionString, _geojsonRoutesContainerName);
+			await DeleteBlob(activityId, BlobType.Geojson);
+			await DeleteBlob(activityId, BlobType.StaticMap);
+		}
+
+		private async Task DeleteBlob(Guid activityId, BlobType type)
+		{
+			if (!_fileExtensions.TryGetValue(type, out var fileExtension))
+			{
+				throw new ArgumentException("Provided blob type is invalid");
+			}
+
+			if (!_containerNames.TryGetValue(type, out var containerName))
+			{
+				throw new ArgumentException("Provided blob type is invalid");
+			}
+
+			var blobName = activityId + fileExtension;
+			var containerClient = new BlobContainerClient(_connectionString, containerName);
 			var blob = containerClient.GetBlobClient(blobName);
 			await blob.DeleteIfExistsAsync();
 		}
